FS-37 Refactor Cache Module
We refactor the `freestyle-cache` module. We remove from the
`freestyle.cache` package any implementation details based on
a Java Concurrent HashMap, and we move those to the `hashmap`
subpackage.

In the `freestyle.cache.implicits`, we leave a simple constructor for
interpreters, based on two steps: a raw implementation, and a natural
transformation.