FS-37 Cache Algebra And Implementations (#70)
* FS-37 Cache Algebra

We add the simple algebra of Cache Operations, to represent
key-value stores. We provide a simple implementation, based
on a java Concurrent Hash Map.

* FS-37 Add Scredis Cache Implementation

We add an implementation of the cache algebra, where the cache is
backed in a Redis data-store. For the connection with the data
store, we need a client library for Scala. We use Scredis.

* FS-37 Kleisli

We replace the previous ad-hoc type ScredisOps with cats.data.Kleisli,
so as to reuse the functions already defined for this type.

* FS-37 Add Example

We add a simple test example, to show the combination of applicative (
(parallel) fragments and the sequential ones.
We also rename the package for the scredis subfolder.x

* FS-37 Replace Scredis by Rediscala

We replace Scredis by Rediscala as our Scala library client for Redis.
Scredis does not seem to be actively maintained.

* FS-37 Refactor Cache Module

We refactor the `freestyle-cache` module. We remove from the
`freestyle.cache` package any implementation details based on
a Java Concurrent HashMap, and we move those to the `hashmap`
subpackage.

In the `freestyle.cache.implicits`, we leave a simple constructor for
interpreters, based on two steps: a raw implementation, and a natural
transformation.

* FS-37 Cache-Refactoring in Redis Module

After the previous refactoring, the Cache Module already separates
between generating a _raw_ format (in the Redis it's Future) and
the Natural Transformation to another carrier `F[_]`.

* FS-37 Add Keys Operation

* Cache HashMap Tests

* FS-37 Resolve Error

The implementation of the `keys` operation in the Hashmap based was
using a `2.12` API operation, unavailable in `2.11` for crossbuild.

* FS-37 Address some Code-Review Comments

* FS-37 Extra comment