FS-87 Use trait-independent type name parameters
We use `trait T[A]` as the ADT root for macro-generated objects, and
so we remove the type alias `T[A] = EffectFoo[A]`.
We use a type parameter `MM[_]` to describe the target functor that a
handler (interpreter) transforms to. Note that, since the scope of a
type parameter is the class Interpreter itself, we can just use one name.
We use a type parameter `LL[_]` for the target functor to which the
Lifter raises a requetst to.