FS-110 Akka-Http Integration
We provide an integration with the akka-http routing DSL,
in line with the integrations for Http4s.

The goal is that, given any expression fs of a type FreeS[F, A],
and if a handler of `F` to a type `G` is available, then we can
put that expression as the argument of a `complete` directive.