FS-177 Make Macro Hygienic
This commit carries a few changes to the `@free` and `@module`
macros, to make them  hygienic as detailed in
http://docs.scala-lang.org/overviews/quasiquotes/hygiene.
The effort helped us to solve a few mistakes we found; but we also
make other changes. In particular:

- In the methods and classes inside the `@free` companion object,
  we avoid using fixed names for the generic types, we use freshTypeName.
- We extract a boolean function `isRequestDef` in the `@free` macro.
- We make the `@module` macro hygienic. This allows us to repair
  an existing bug, by which only the `F[_]` could be used as argument
  in the `@module`-annotated traits.

As a bit of modularity, we separate the Macro-annotations to a
distinct file from those of the macro implementations.

To increase the code coverage, we also add some tests for the
errors reported by wrong macro usage.