FS-234 Unify Seq and Par
We remove the distinction between Parallel and Sequential operations
at the level of the `@free` annotation for traits.
In addition to `OpSeq` and `OpPar`, the `@free` trait includes the
high-kind `FS[_]`, defined as an alias for `FreeS.Par[F, _].

In a `@free`-annotated trait, all requests (abstract methods) must
have the type FS[Res]. We leave the OpSeq and OpPar types, but only
for those methods that include  an implementation, based on monadic
or applicative operations of the FreeS types.