FS-237 Adjust Docs to new Module Macro.
As an effect of the recent changes in `@module`, now it is not
possible to have in a `tut` session an `@free` algebra and a `@module`
that uses it. For this reason, we need to copy those algebras (or
modules) to auxiliary scala files, and import them in the `tut` books.