FS-73-1 Reimplement @free macro in Scalameta.
We reimplement the `@free` macro using Scalameta. The new
implementation

There are several cases in which Scalameta quasiquotes are less
flexible than scala.reflect macros. For instance, Scalameta does not
provide a `cq""` quasiquote for match cases, so we need either to
build them ourselves, or to build a complete `match` to get it.

In the same vein, taking a type name `F` and writing into a quasiquote
`F[_]` is not as flexible unless we can align the actual type of F.
Also, we can not unquote the integer variable $indexValue.

On the other hand, because Scalameta has a representation of Scala AST
that is completely independent of the context, we now can make our
class `Request` into a top level (`package`), and we can extract
some Scalameta utilities, independent of freestyle, to another file.

Since the package `freestyle` has too many members, we have moved the
macro implementation into an inner `macroimpl` package.