FS-495 Avoid flattening lists of parameters (#496)
FS-496 Handle Generic Implicit parameters

Previously, the `@free` macro was flattening the parameters of each request method, i.e., it was joining all the lists of parameters into a single list. This has obvious problem of being unintuitive, but it also made it difficult to handle `implicit` parameters and context bounds on type parameters: in `Scala`, all parameters in each list have to be either `implicit` or non implicit. 
 
We modify the `@free` macro, so as to ensure that we do not flatten the lists of parameters. We add some tests to check for such cases.