Very minor grammar changes
You could use either both, depending on your requirements, and the
usage would be quite similar.

You could use either, or both, depending on your requirements, and the
usage would be quite similar.

If that is not accurate, just take out the or both and change to:

You could use either, depending on your requirements, and the usage
would be quite similar.