FS-517 Isues with StackSafe Expansion.
There are some issues in `@tagless` when using stacksafe, which are
due to the fact that the tagless is more flexible than the @free
macro.

To avoid these issues, we have separated a `@stacksafe` annotation for
the `@tagless` annotated traits. Without it, it does not generate the
Stacksafe traits and object. (the free for the algebra)