FS-532 StackSafe as an argument of Tagless.
Having @stacksafe as a separate annotation (or mod) was giving
problems with IntelliJ Idea, so we cannot use that any more.

Instead, we add an optional argument to tagless, the argument
stacksafe, following the pattern shown here.
http://scalameta.org/paradise/#HowdoIpassanargumenttothemacroannotation?

We replace the code in all of the tests and the documentations we have
found so far.